Refactored Orbits package to be more easily utilized.
Orbits.new_orbit() is now the official gateway into the library, returning an Orbit object that has functions for giving both the cartesian and polar coordinates of a satellite on a defined orbit. The defining orbital elements are contained in a another struct in the Orbit object.

Still need to do testing to fully characterize the performance and convergence behavior of my Newton-Raphson implementation, but it works decently well for now.

The function giving the polar angle of the orbit over time is relatively jagged right now, however. Will be testing this some more.